feat: access adapter removal keccak256
Cast usize pointers to u32 before to_le_bytes() to produce 4-byte
arrays compatible with BLOCK_SIZE=4 memory configuration.

+++ extensions/keccak256/circuit/src/xorin/tests.rs
@@ -122,12 +122,12 @@ fn set_and_execute<RA: Arena, E: PreflightExecutor<F, RA>>(
         tester.write(2, input_ptr + 4 * i, input_chunk);
     }
 
-    tester.write(1, rd, buffer_ptr.to_le_bytes().map(F::from_canonical_u8));
-    tester.write(1, rs1, input_ptr.to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rd, (buffer_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rs1, (input_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
     tester.write(
         1,
         rs2,
-        buffer_length.to_le_bytes().map(F::from_canonical_u8),
+        (buffer_length as u32).to_le_bytes().map(F::from_canonical_u8),
     );
     tester.execute(
         executor,
